requirements.txt
google-generativeai>=0.2.0 
tkinter
time
